# pedi-agent-stack/openapi/pedi_agents.openapi.yaml
openapi: 3.0.3
info:
  title: PediScreen Agents API
  version: "1.0.0"
  description: API spec for Intake, Embedding, VisionQA, Temporal, MedGemma, Safety, Jobs, and Webhooks
servers:
  - url: http://localhost:7000
paths:
...
  /v1/jobs/{job_id}:
    get:
      summary: Job status
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  result: { type: object }
  /v1/webhooks/register:
    post:
      summary: Register webhook for clinic notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string }
                secret_name: { type: string }
      responses:
        '201':
          description: webhook registered
...